Fix tests

+++ db/schema.rb
@@ -1420,7 +1420,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_03_10_170809) do
     t.datetime "created_at", null: false
     t.string "status", null: false
     t.datetime "updated_at", null: false
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.index ["user_id"], name: "index_tickets_claim_wca_id_on_user_id"
   end
